fix(clerk-js,shared): Refactor and move loadScript into shared
