+++ packages/clerk-js/src/utils/index.ts
@@ -20,7 +20,6 @@ export * from './queryStateParams';
 export * from './querystring';
 export * from './runtime';
 export * from './safeLock';
-export * from './script';
 export * from './url';
 export * from './web3';
 export * from './windowNavigate';
